+++ spotifyUtils/seedData.js
@@ -1,0 +1,24 @@
+//create room
+const roomData = {
+    roomName: 'TestRoom2',
+    paylistId: '123456abc',
+    passcode: 123,
+    hostId: 1,
+    accessToken: '123ABC',
+    refreshToken: '345DEF',
+    expirationTime: 'now'
+  };
+  const song1 = {
+    name: 'free bird',
+    id: '123ABC',
+    userId: 'Bob',
+    upvote: false,
+    downvote: false
+  };
+  const song2 = {
+    name: 'piano man',
+    id: '345DFE',
+    userId: 'Billy',
+    upvote: false,
+    downvote: false
+  };
